ai_worker/pyproject.toml
[project]
name = "ai_worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard]>=0.128.8",
    "instructor>=1.14.5",
    "openai>=2.29.0",
    "python-dotenv>=1.2.1",
    "instaloader>=4.14",
]

[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint]
select = ["E", "F", "I"]

[dependency-groups]
dev = [
    "ruff>=0.15.8",
]
